flask-backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
Flask-Bcrypt
psycopg2-binary
